IBX-10671: Merge Icon and IconCustom components
