--- src/lib/Twig/Components/AbstractIcon.php
@@ -1,41 +1,0 @@
-<?php
-
-/**
- * @copyright Copyright (C) Ibexa AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- */
-declare(strict_types=1);
-
-namespace Ibexa\DesignSystemTwig\Twig\Components;
-
-use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\UX\TwigComponent\Attribute\PreMount;
-
-abstract class AbstractIcon
-{
-    public string $size = 'medium';
-
-    public string $path = '';
-
-    /**
-     * @param array<string, mixed> $props
-     *
-     * @return array<string, mixed>
-     */
-    #[PreMount]
-    public function validate(array $props): array
-    {
-        $resolver = new OptionsResolver();
-        $resolver->setIgnoreUndefined();
-        $resolver
-            ->define('size')
-            ->allowedValues('tiny', 'tiny-small', 'small', 'small-medium', 'medium', 'medium-large', 'large', 'extra-large', 'large-huge', 'huge')
-            ->default('medium');
-
-        $this->configurePropsResolver($resolver);
-
-        return $resolver->resolve($props) + $props;
-    }
-
-    abstract protected function configurePropsResolver(OptionsResolver $resolver): void;
-}
